+++ src/main/java/com/github/system/base/entity/SysWebhook.java
@@ -1,5 +1,7 @@
 package com.github.system.base.entity;
 
+import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -11,10 +13,10 @@ import lombok.Data;
 public class SysWebhook {
 
     @ApiModelProperty("主键")
-    private Integer id;
+    private Long id;
 
     @ApiModelProperty("用户id")
-    private Integer userId;
+    private Long userId;
 
     @ApiModelProperty("推送类型")
     private String type;
