+++ src/main/java/com/github/system/base/service/WebhookService.java
@@ -8,7 +8,7 @@ import com.github.system.base.entity.SysWebhook;
 import java.util.List;
 
 public interface WebhookService extends IService<SysWebhook> {
-    List<SysWebhook> getUserWebhook(int userId);
+    List<SysWebhook> getUserWebhook(long userId);
 
     AjaxResult checkWebhook(PushData<?> pushData);
 }
